requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.11.2
Flask==2.3.3
Flask_Cors==4.0.0
flask-restx==1.3.0
Werkzeug==2.3.7
gdown==5.2.0
matplotlib==3.8.4
numpy==2.0.1
pandas==2.3.0
python-dotenv==1.1.1
Requests==2.32.4
scikit_learn==1.4.2
seaborn==0.13.2
supabase==2.16.0
tqdm==4.67.1
upstash_redis==1.4.0
